# Optimize path compression of union-find