@QuarticCat 2023-04-03
The tests `pattern::tests::simple_match` and `explain::tests::simple_explain` contain hard-coded paths to the building target dir. If someone changes their target dir settings (like me), these two tests will break. Would you like to fix it? I can add a commit to this thread.