@QuarticCat 2023-04-03
After I updated my benchmarking environment, the speedup is negligible. So I gonna close this PR. The bench script can be found here https://github.com/QuarticCat/dotfiles/blob/main/zsh/scripts/bench-mode.zsh